chore: renamed extract.rs to install.rs
