+++ src-tauri/src/install.rs
